requirements2.txt
pyyaml
tqdm
numpy
easydict==1.9.0
scikit-image
scikit-learn
joblib
matplotlib
pandas
albumentations==0.5.2
hydra-core==1.1.0
pytorch-lightning==1.2.9
tabulate
kornia==0.5.0
webdataset
packaging
wget